指摘箇所の修正

+++ app/controllers/items_controller.rb
@@ -41,6 +41,9 @@ class ItemsController < ApplicationController
     end
 
     def destroy
+      if @item.user_id != current_user.id
+        redirect_to action: :index
+      end
       item = Item.find(params[:id])
       item.destroy
       redirect_to root_path
